+++ decidim-core/db/migrate/20181029121244_add_welcome_notification.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class AddWelcomeNotification < ActiveRecord::Migration[5.2]
+  def change
+    change_table :decidim_organizations do |t|
+      t.boolean :send_welcome_notification, null: false, default: false
+      t.jsonb :welcome_notification_subject, null: true
+      t.jsonb :welcome_notification_body, null: true
+    end
+  end
+end